fix gradle external files caching problem
